refactor: update package structure for data transformers to improve organization
